fix: trim padding from length calculation
Padding ("=" characters) decodes as 0, so it doesn't affect the total for
the block it's in. However, padding that overflows into extra blocks
would append extra null bytes. Since the `catch` block only runs as a
fallback when `atob` throws, which could be due to faulty padding, the
padding shouldn't be considered in the final length calculation. Mutating
the argument achieves this when done outside the loop's scope.